batch file to create dpi images for android.
You will need image magick in your path.

http://www.imagemagick.org/script/index.php

I recommend putting this batch file in your path also. To use the batch file, give it 2 parameters which are the image file name and optionally a base scaling level. The base scaling level will be used to control the amount of upscaling. If you have full control over your artwork I recommend designing the base graphic at the highest level resolution and passing the batch file a parameter of 1 for the scaleLevel so that everything is nicely downsized instead of upsized.